Codes/AIEditor/requirements.txt
asgiref==3.8.1
cryptography==42.0.8
Django==5.0.6
djangorestframework==3.15.2
djangorestframework_simplejwt==5.3.1
erniebot_agent==0.5.2
fitz==0.0.1.dev2
fonttools==4.53.0
langchain==0.2.9
langchain_community==0.2.7
numpy==2.0.0
opencv_contrib_python==4.6.0.66
opencv_python==4.6.0.66
opencv_python_headless==4.10.0.84
paddleocr==2.7.3
paddleocr.egg==info
paddlepaddle_gpu==2.6.1.post120
PyMySQL==1.1.1
python_alipay_sdk==3.3.0
Requests==2.32.3
scikit_learn==1.5.1
